requirements.txt
opencv-python<4.10
numpy<2
ultralytics
torch
torchvision
torchaudio
matplotlib
pandas
pyyaml
tqdm
jupyter
ipython
tabulate